Moved the type change detection to a separate class
